# Perf: create a tracker per module and use indexes to improve performance