--- a/src/coverlet.core/Instrumentation/Instrumenter.cs
+++ b/src/coverlet.core/Instrumentation/Instrumenter.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using System.Linq;
@@ -7,7 +8,6 @@ using System.Reflection;
 using Coverlet.Core.Attributes;
 using Coverlet.Core.Helpers;
 using Coverlet.Core.Symbols;
-using Coverlet.Tracker;
 
 using Mono.Cecil;
 using Mono.Cecil.Cil;
@@ -22,8 +22,11 @@ namespace Coverlet.Core.Instrumentation
         private readonly string[] _excludeFilters;
         private readonly string[] _includeFilters;
         private readonly string[] _excludedFiles;
-        private readonly static Lazy<MethodInfo> _markExecutedMethodLoader = new Lazy<MethodInfo>(GetMarkExecutedMethod);
         private InstrumenterResult _result;
+        private FieldDefinition _customModuleTrackerHitsArray;
+        private FieldDefinition _customModuleTrackerHitsFilePath;
+        private ILProcessor _customModuleTrackerClassConstructorIl;
+        private MethodReference _cachedInterlockedIncMethod;
 
         public Instrumenter(string module, string identifier, string[] excludeFilters, string[] includeFilters, string[] excludedFiles)
         {
@@ -51,7 +54,6 @@ namespace Coverlet.Core.Instrumentation
             };
 
             InstrumentModule();
-            InstrumentationHelper.CopyCoverletDependency(_module);
 
             return _result;
         }
@@ -67,18 +69,107 @@ namespace Coverlet.Core.Instrumentation
                 using (var module = ModuleDefinition.ReadModule(stream, parameters))
                 {
                     var types = module.GetTypes();
+                    AddCustomModuleTrackerToModule(module);
+
                     foreach (TypeDefinition type in types)
                     {
                         var actualType = type.DeclaringType ?? type;
                         if (!actualType.CustomAttributes.Any(IsExcludeAttribute)
+                            && actualType.Namespace != "Coverlet.Core.Instrumentation.Tracker"
                             && !InstrumentationHelper.IsTypeExcluded(_module, actualType.FullName, _excludeFilters)
                             && InstrumentationHelper.IsTypeIncluded(_module, actualType.FullName, _includeFilters))
                             InstrumentType(type);
                     }
 
+                    // Fixup the custom tracker class constructor, according to all instrumented types
+                    Instruction lastInstr = _customModuleTrackerClassConstructorIl.Body.Instructions.Last();
+                    _customModuleTrackerClassConstructorIl.InsertBefore(lastInstr, Instruction.Create(OpCodes.Ldc_I4, _result.HitCandidates.Count));
+                    _customModuleTrackerClassConstructorIl.InsertBefore(lastInstr, Instruction.Create(OpCodes.Newarr, module.TypeSystem.Int32));
+                    _customModuleTrackerClassConstructorIl.InsertBefore(lastInstr, Instruction.Create(OpCodes.Stsfld, _customModuleTrackerHitsArray));
+                    _customModuleTrackerClassConstructorIl.InsertBefore(lastInstr, Instruction.Create(OpCodes.Ldstr, _result.HitsFilePath));
+                    _customModuleTrackerClassConstructorIl.InsertBefore(lastInstr, Instruction.Create(OpCodes.Stsfld, _customModuleTrackerHitsFilePath));
+
                     module.Write(stream);
                 }
             }
+        }
+
+        private void AddCustomModuleTrackerToModule(ModuleDefinition module)
+        {
+            using (AssemblyDefinition xad = AssemblyDefinition.ReadAssembly(typeof(Instrumenter).Assembly.Location))
+            {
+                TypeDefinition moduleTrackerTemplate = xad.MainModule.GetType(
+                    "Coverlet.Core.Instrumentation", nameof(ModuleTrackerTemplate));
+
+                TypeDefinition customTrackerTypeDef = new TypeDefinition(
+                    "Coverlet.Core.Instrumentation.Tracker", Path.GetFileNameWithoutExtension(module.Name) + "_" + _identifier, moduleTrackerTemplate.Attributes);
+
+                customTrackerTypeDef.BaseType = module.TypeSystem.Object;
+                foreach (FieldDefinition fieldDef in moduleTrackerTemplate.Fields)
+                {
+                    var fieldClone = new FieldDefinition(fieldDef.Name, fieldDef.Attributes, fieldDef.FieldType);
+                    customTrackerTypeDef.Fields.Add(fieldClone);
+
+                    if (fieldClone.Name == "HitsArray")
+                        _customModuleTrackerHitsArray = fieldClone;
+                    else if (fieldClone.Name == "HitsFilePath")
+                        _customModuleTrackerHitsFilePath = fieldClone;
+                }
+
+                foreach (MethodDefinition methodDef in moduleTrackerTemplate.Methods)
+                {
+                    MethodDefinition methodOnCustomType = new MethodDefinition(methodDef.Name, methodDef.Attributes, methodDef.ReturnType);
+
+                    foreach (var variable in methodDef.Body.Variables)
+                    {
+                        methodOnCustomType.Body.Variables.Add(new VariableDefinition(module.ImportReference(variable.VariableType)));
+                    }
+
+                    methodOnCustomType.Body.InitLocals = methodDef.Body.InitLocals;
+
+                    ILProcessor ilProcessor = methodOnCustomType.Body.GetILProcessor();
+                    if (methodDef.Name == ".cctor")
+                        _customModuleTrackerClassConstructorIl = ilProcessor;
+
+                    foreach (Instruction instr in methodDef.Body.Instructions)
+                    {
+                        if (instr.Operand is MethodReference methodReference)
+                        {
+                            if (!methodReference.FullName.Contains(moduleTrackerTemplate.Namespace))
+                            {
+                                // External method references, just import then
+                                instr.Operand = module.ImportReference(methodReference);
+                            }
+                            else
+                            {
+                                // Move to the custom type
+                                instr.Operand = new MethodReference(
+                                    methodReference.Name, methodReference.ReturnType, customTrackerTypeDef);
+                            }
+                        }
+                        else if (instr.Operand is FieldReference fieldReference)
+                        {
+                            instr.Operand = customTrackerTypeDef.Fields.Single(fd => fd.Name == fieldReference.Name);
+                        }
+                        else if (instr.Operand is TypeReference typeReference)
+                        {
+                            instr.Operand = module.ImportReference(typeReference);
+                        }
+
+                        ilProcessor.Append(instr);
+                    }
+
+                    foreach (var handler in methodDef.Body.ExceptionHandlers)
+                        methodOnCustomType.Body.ExceptionHandlers.Add(handler);
+
+                    customTrackerTypeDef.Methods.Add(methodOnCustomType);
+                }
+
+                module.Types.Add(customTrackerTypeDef);
+            }
+
+            Debug.Assert(_customModuleTrackerHitsArray != null);
+            Debug.Assert(_customModuleTrackerClassConstructorIl != null);
         }
 
         private void InstrumentType(TypeDefinition type)
@@ -143,7 +234,7 @@ namespace Coverlet.Core.Instrumentation
                     foreach (ExceptionHandler handler in processor.Body.ExceptionHandlers)
                         ReplaceExceptionHandlerBoundary(handler, instruction, target);
 
-                    index += 3;
+                    index += 5;
                 }
 
                 foreach (var _branchTarget in targetedBranchPoints)
@@ -164,7 +255,7 @@ namespace Coverlet.Core.Instrumentation
                     foreach (ExceptionHandler handler in processor.Body.ExceptionHandlers)
                         ReplaceExceptionHandlerBoundary(handler, instruction, target);
 
-                    index += 3;
+                    index += 5;
                 }
 
                 index++;
@@ -188,17 +279,10 @@ namespace Coverlet.Core.Instrumentation
                     document.Lines.Add(i, new Line { Number = i, Class = method.DeclaringType.FullName, Method = method.FullName });
             }
 
-            string marker = $"L,{document.Index},{sequencePoint.StartLine},{sequencePoint.EndLine}";
+            var entry = (false, document.Index, sequencePoint.StartLine, sequencePoint.EndLine);
+            _result.HitCandidates.Add(entry);
 
-            var pathInstr = Instruction.Create(OpCodes.Ldstr, _result.HitsFilePath);
-            var markInstr = Instruction.Create(OpCodes.Ldstr, marker);
-            var callInstr = Instruction.Create(OpCodes.Call, processor.Body.Method.Module.ImportReference(_markExecutedMethodLoader.Value));
-
-            processor.InsertBefore(instruction, callInstr);
-            processor.InsertBefore(callInstr, markInstr);
-            processor.InsertBefore(markInstr, pathInstr);
-
-            return pathInstr;
+            return AddInstrumentationInstructions(method, processor, instruction, _result.HitCandidates.Count);
         }
 
         private Instruction AddInstrumentationCode(MethodDefinition method, ILProcessor processor, Instruction instruction, BranchPoint branchPoint)
@@ -225,17 +309,34 @@ namespace Coverlet.Core.Instrumentation
                     }
                 );
 
-            string marker = $"B,{document.Index},{branchPoint.StartLine},{branchPoint.Ordinal}";
+            var entry = (true, document.Index, branchPoint.StartLine, (int)branchPoint.Ordinal);
+            _result.HitCandidates.Add(entry);
 
-            var pathInstr = Instruction.Create(OpCodes.Ldstr, _result.HitsFilePath);
-            var markInstr = Instruction.Create(OpCodes.Ldstr, marker);
-            var callInstr = Instruction.Create(OpCodes.Call, processor.Body.Method.Module.ImportReference(_markExecutedMethodLoader.Value));
+            return AddInstrumentationInstructions(method, processor, instruction, _result.HitCandidates.Count);
+        }
 
-            processor.InsertBefore(instruction, callInstr);
-            processor.InsertBefore(callInstr, markInstr);
-            processor.InsertBefore(markInstr, pathInstr);
+        private Instruction AddInstrumentationInstructions(MethodDefinition method, ILProcessor processor, Instruction instruction, int hitEntryIndex)
+        {
+            if (_cachedInterlockedIncMethod == null)
+            {
+                _cachedInterlockedIncMethod = new MethodReference(
+                    "Increment", method.Module.TypeSystem.Int32, method.Module.ImportReference(typeof(System.Threading.Interlocked)));
+                _cachedInterlockedIncMethod.Parameters.Add(new ParameterDefinition(new ByReferenceType(method.Module.TypeSystem.Int32)));
+            }
 
-            return pathInstr;
+            var sfldInstr = Instruction.Create(OpCodes.Ldsfld, _customModuleTrackerHitsArray);
+            var indxInstr = Instruction.Create(OpCodes.Ldc_I4, hitEntryIndex);
+            var arefInstr = Instruction.Create(OpCodes.Ldelema, method.Module.TypeSystem.Int32);
+            var callInstr = Instruction.Create(OpCodes.Call, _cachedInterlockedIncMethod);
+            var popInstr = Instruction.Create(OpCodes.Pop);
+
+            processor.InsertBefore(instruction, popInstr);
+            processor.InsertBefore(popInstr, callInstr);
+            processor.InsertBefore(callInstr, arefInstr);
+            processor.InsertBefore(arefInstr, indxInstr);
+            processor.InsertBefore(indxInstr, sfldInstr);
+
+            return sfldInstr;
         }
 
         private static void ReplaceInstructionTarget(Instruction instruction, Instruction oldTarget, Instruction newTarget)
@@ -300,11 +401,6 @@ namespace Coverlet.Core.Instrumentation
             {
                 return null;
             }
-        }
-
-        private static MethodInfo GetMarkExecutedMethod()
-        {
-            return typeof(CoverageTracker).GetMethod(nameof(CoverageTracker.MarkExecuted));
         }
     }
 }