--- a/src/coverlet.tracker/Properties/AssemblyInfo.cs
+++ b/src/coverlet.tracker/Properties/AssemblyInfo.cs
@@ -1,1 +1,0 @@
-[assembly: System.Reflection.AssemblyKeyFileAttribute("coverlet.tracker.snk")]