--- a/src/coverlet.tracker/coverlet.tracker.csproj
+++ b/src/coverlet.tracker/coverlet.tracker.csproj
@@ -1,7 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-</Project>