--- a/test/coverlet.core.performancetest/PerformanceTest.cs
+++ b/test/coverlet.core.performancetest/PerformanceTest.cs
@@ -15,8 +15,8 @@ namespace coverlet.core.performancetest
     public class PerformanceTest
     {
         [Theory(/*Skip = "Only enabled when explicitly testing performance."*/)]
-        [InlineData(150)]
-        // [InlineData(20_000)]
+        // [InlineData(150)]
+        [InlineData(20_000)]
         public void TestPerformance(int iterations)
         {
             var big = new BigClass();