… coverlet.sln
… src/coverlet.tracker/Extensions/DictionaryExtensions.cs
… src/coverlet.tracker/coverlet.tracker.snk
… test/coverlet.core.tests/Helpers/InstrumentationHelperTests.cs
… test/coverlet.core.tests/Instrumentation/ModuleTrackerTemplateTests.cs