@pjanotti 2018-08-22
@tonerdo I know that you just merged #172 but I got this PR ready :grinning: 

I will be slow to answer any Q/comments for a few days, so no rush with this one. It will be great if other contributors could spin this PR against their projects just to be sure that there isn't some corner case missed here.

/cc @tarekgh @petli