@petli 2018-08-24
Testing this PR against one of the bigger unit test projects in the code I'm working on at the day job the performance improvements are very good:

Test type | Test execution time | Total time to run dotnet test --no-build
----------|--------------------|--------------
Without coverage | 30s | 31s
coverlet 2.2.1   | 133s | 144s
PR changes       | 34s | 37s

Impressive, @pjanotti!